request: move CqlRequestKind enum to request module
It should not belong to errors module.
