+++ scylla-cql/src/frame/request/mod.rs
@@ -24,6 +24,37 @@ use self::batch::BatchStatement;
 use super::types::SerialConsistency;
 use super::TryFromPrimitiveError;
 
+/// Possible requests sent by the client.
+#[derive(Debug, Copy, Clone)]
+#[non_exhaustive]
+pub enum CqlRequestKind {
+    Startup,
+    AuthResponse,
+    Options,
+    Query,
+    Prepare,
+    Execute,
+    Batch,
+    Register,
+}
+
+impl std::fmt::Display for CqlRequestKind {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let kind_str = match self {
+            CqlRequestKind::Startup => "STARTUP",
+            CqlRequestKind::AuthResponse => "AUTH_RESPONSE",
+            CqlRequestKind::Options => "OPTIONS",
+            CqlRequestKind::Query => "QUERY",
+            CqlRequestKind::Prepare => "PREPARE",
+            CqlRequestKind::Execute => "EXECUTE",
+            CqlRequestKind::Batch => "BATCH",
+            CqlRequestKind::Register => "REGISTER",
+        };
+
+        f.write_str(kind_str)
+    }
+}
+
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
 #[repr(u8)]
 pub enum RequestOpcode {
